changed tabs views to separate folder
