+++ frontend/src/components/shape/tab/PrefixTab.vue
@@ -37,7 +37,7 @@
   // // import transformString from '@/mixins/transformString';
   import { validateAbsoluteIRI, validatePrefix } from '@/mixins/validateShape';
   import { ErrorMessage, Field } from 'vee-validate';
-  import shapeConfig from '../../../../resources/shapes/config.json';
+  import shapeConfig from '../../../../../resources/shapes/config.json';
   //
   export default defineComponent({
     components: { ErrorMessage, Field },
